app/requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
redis==5.0.8
rq==1.16.2
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
alembic==1.13.2
openai==1.52.2
streamlit==1.39.0
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0